Replacing nanos6 interface with the ALPI tasking interface
This commit replaces the nanos6 interface with the ALPI tasking interface. This
latter is more generic and can be implemented by any tasking runtime system.
Modern versions of the Nanos6 and nOS-V runtime libraries should implement and
expose it

The support for the ALPI interface is added in OmpSs-2 2023.11. Previous versions
of OmpSs-2 are not compatible. Concretely, the support for the interface in the
runtime systems' side is added in Nanos6 4.0 and nOS-V 2.0